Update to match new DOM from Amazon
Numeraous changes happened to the Kindle review
page. It took care of some of the friction of
eralier implementation.

It is no longer the case that all highlights are
loaded in one page, lazily. We now have per-book
notes. This greatly simplifies the logic.

Also we ensured that the listener is only bound
once.

There are a few other refactorings, but it's a
very nice simplification.